move meters verification
