+++ src/app/upload-data/data-setup/file-setup/file-setup.component.html
@@ -18,6 +18,11 @@
                     <i class="fa fa-gauge-high"></i>Map Facility Meters</a>
             </li>
             <li class="nav-item">
+                <a class="nav-link" routerLink="manage-meters"
+                [routerLinkActive]="['active']">
+                    <i class="fa fa-gauge-high"></i>Verify Meters</a>
+            </li>
+            <li class="nav-item">
                 <a class="nav-link" routerLink="set-facility-predictors"
                 [routerLinkActive]="['active']">
                     <i class="fa fa-boxes-stacked"></i>Map Facility Predictors</a>
